Integrate Google Test and add for basic operations.
Integrate the Google C++ Testing Framework.

Unit tests and implementations for hashing and equality of Arc and State
classes. This will be needed to implement hashing of states for storing
them in a hashtable.
Unit tests and implementations for Fsa construction from a vector of arcs
and checking if a word is accepted by the automaton.